Implemented input data and result data setting for test 5 in both screen, also improved Formula method  to support longer data and also did modification in test 5 UI for dynamic rows 7 columns

+++ VALDOC/app/src/main/java/com/project/valdoc/RDAV5UserEntryActivity.java
@@ -127,7 +127,7 @@ public class RDAV5UserEntryActivity extends AppCompatActivity {
         }
 
         //Receiving Result Data from Bundle
-        HashMap<Integer, Integer> resultHashMap = (HashMap<Integer, Integer>) getIntent().getSerializableExtra("ResultData");
+        HashMap<Integer, Long> resultHashMap = (HashMap<Integer, Long>) getIntent().getSerializableExtra("ResultData");
         for (Map.Entry n : resultHashMap.entrySet()) {
             Log.v(TAG, " Result: " + n.getKey() + " " + n.getValue());
         }
